Move RS_NormalizeAll to Raster Editors; Update docs
