+++ common/src/test/java/org/apache/sedona/common/raster/RasterEditorsTest.java
@@ -18,11 +18,13 @@
  */
 package org.apache.sedona.common.raster;
 
+import org.apache.sedona.common.utils.RasterUtils;
 import org.geotools.coverage.grid.GridCoverage2D;
 import org.junit.Test;
 import org.opengis.referencing.FactoryException;
 import org.opengis.referencing.operation.TransformException;
 
+import java.awt.image.DataBuffer;
 import java.io.IOException;
 import java.util.Arrays;
 
@@ -264,6 +266,131 @@ public class RasterEditorsTest extends RasterTestBase {
         GridCoverage2D raster = RasterConstructors.makeEmptyRaster(1, "d", 3, 3, 0, 0, 2, -2, 0, 0, 0);
         GridCoverage2D refRaster = RasterConstructors.makeEmptyRaster(2, "d", 5, 5, 1, -1, 1.2, -1.2, 0, 0, 4326);
         assertThrows("Provided input raster and reference raster have different SRIDs", IllegalArgumentException.class, () -> RasterEditors.resample(raster, refRaster, false, null));
+    }
+
+    @Test
+    public void testNormalizeAll() throws FactoryException {
+        GridCoverage2D raster1 = RasterConstructors.makeEmptyRaster(2, 4, 4, 0, 0, 1);
+        GridCoverage2D raster2 = RasterConstructors.makeEmptyRaster(2, 4, 4, 0, 0, 1);
+        GridCoverage2D raster3 = RasterConstructors.makeEmptyRaster(2, "I", 4, 4, 0, 0, 1);
+        GridCoverage2D raster4 = RasterConstructors.makeEmptyRaster(2, 4, 4, 0, 0, 1);
+        GridCoverage2D raster5 = RasterConstructors.makeEmptyRaster(2, 4, 4, 0, 0, 1);
+
+        for (int band = 1; band <= 2; band++) {
+            double[] bandValues1 = new double[4 * 4];
+            double[] bandValues2 = new double[4 * 4];
+            double[] bandValues3 = {1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16};
+            double[] bandValues4 = {1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,0};
+            double[] bandValues5 = new double[4 * 4];
+            for (int i = 0; i < bandValues1.length; i++) {
+                bandValues1[i] = (i) * band;
+                bandValues2[i] = (1) * (band-1);
+                bandValues5[i] = i + ((band-1)*15);
+            }
+            raster1 = MapAlgebra.addBandFromArray(raster1, bandValues1, band);
+            raster2 = MapAlgebra.addBandFromArray(raster2, bandValues2, band);
+            raster3 = MapAlgebra.addBandFromArray(raster3, bandValues3, band);
+            raster4 = MapAlgebra.addBandFromArray(raster4, bandValues4, band);
+            raster4 = RasterBandEditors.setBandNoDataValue(raster4, band, 0.0);
+            raster5 = MapAlgebra.addBandFromArray(raster5, bandValues5, band);
+        }
+        raster3 = RasterBandEditors.setBandNoDataValue(raster3, 1, 16.0);
+        raster3 = RasterBandEditors.setBandNoDataValue(raster3, 2, 1.0);
+
+        GridCoverage2D normalizedRaster1 = RasterEditors.normalizeAll(raster1, 0, 255, false, -9999.0);
+        GridCoverage2D normalizedRaster2 = RasterEditors.normalizeAll(raster1, 256d, 511d, false, -9999.0);
+        GridCoverage2D normalizedRaster3 = RasterEditors.normalizeAll(raster2);
+        GridCoverage2D normalizedRaster4 = RasterEditors.normalizeAll(raster3, 0, 255, true, 95.0);
+        GridCoverage2D normalizedRaster5 = RasterEditors.normalizeAll(raster4, 0, 255, true, 255.0);
+        GridCoverage2D normalizedRaster6 = RasterEditors.normalizeAll(raster5, 0.0, 255.0, -9999.0, 0.0, 30.0);
+        GridCoverage2D normalizedRaster7 = RasterEditors.normalizeAll(raster5, 0, 255, false, -9999.0);
+        GridCoverage2D normalizedRaster8 = RasterEditors.normalizeAll(raster3, 0, 255);
+        GridCoverage2D normalizedRaster10 = RasterEditors.normalizeAll(raster3, 0, 255, false);
+
+        double[] expected1 = {0.0, 17.0, 34.0, 51.0, 68.0, 85.0, 102.0, 119.0, 136.0, 153.0, 170.0, 187.0, 204.0, 221.0, 238.0, 255.0};
+        double[] expected2 = {256.0, 273.0, 290.0, 307.0, 324.0, 341.0, 358.0, 375.0, 392.0, 409.0, 426.0, 443.0, 460.0, 477.0, 494.0, 511.0};
+        double[] expected3 = {0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0};
+        double[] expected4 = {0.0, 17.0, 34.0, 51.0, 68.0, 85.0, 102.0, 119.0, 136.0, 153.0, 170.0, 187.0, 204.0, 221.0, 238.0, 95.0};
+        double[] expected5 = {95.0, 17.0, 34.0, 51.0, 68.0, 85.0, 102.0, 119.0, 136.0, 153.0, 170.0, 187.0, 204.0, 221.0, 238.0, 255.0};
+        double[] expected6 = {0.0, 18.214285714285715, 36.42857142857143, 54.642857142857146, 72.85714285714286, 91.07142857142857, 109.28571428571429, 127.5, 145.71428571428572, 163.92857142857142, 182.14285714285714, 200.35714285714286, 218.57142857142858, 236.78571428571428, 255.0, 255.0};
+        double[] expected7 = {0.0, 16.0, 33.0, 50.0, 67.0, 84.0, 101.0, 118.0, 135.0, 152.0, 169.0, 186.0, 203.0, 220.0, 237.0, 255.0};
+        double[] expected8 = {255.0, 16.0, 33.0, 50.0, 67.0, 84.0, 101.0, 118.0, 135.0, 152.0, 169.0, 186.0, 203.0, 220.0, 237.0, 254.0};
+        double[] expected9 = {0.0, 18.0, 36.0, 54.0, 72.0, 90.0, 108.0, 127.0, 145.0, 163.0, 181.0, 199.0, 217.0, 235.0, 254.0, 255.0};
+        double[] expected10 = {255.0, 0.0, 18.0, 36.0, 54.0, 72.0, 90.0, 108.0, 127.0, 145.0, 163.0, 181.0, 199.0, 217.0, 235.0, 254.0};
+
+        // Step 3: Validate the results for each band
+        for (int band = 1; band <= 2; band++) {
+            double[] normalizedBand1 = MapAlgebra.bandAsArray(normalizedRaster1, band);
+            double[] normalizedBand2 = MapAlgebra.bandAsArray(normalizedRaster2, band);
+            double[] normalizedBand5 = MapAlgebra.bandAsArray(normalizedRaster5, band);
+            double[] normalizedBand6 = MapAlgebra.bandAsArray(normalizedRaster6, band);
+            double[] normalizedBand7 = MapAlgebra.bandAsArray(normalizedRaster7, band);
+            double normalizedMin6 = Arrays.stream(normalizedBand6).min().getAsDouble();
+            double normalizedMax6 = Arrays.stream(normalizedBand6).max().getAsDouble();
+
+            assertEquals(Arrays.toString(expected1), Arrays.toString(normalizedBand1));
+            assertEquals(Arrays.toString(expected2), Arrays.toString(normalizedBand2));
+            assertEquals(Arrays.toString(expected6), Arrays.toString(normalizedBand5));
+            assertEquals(Arrays.toString(expected1), Arrays.toString(normalizedBand7));
+
+            assertEquals(0+((band-1)*127.5), normalizedMin6, 0.01d);
+            assertEquals(127.5+((band-1)*127.5), normalizedMax6, 0.01d);
+        }
+
+        assertEquals(95.0, RasterUtils.getNoDataValue(normalizedRaster4.getSampleDimension(0)), 0.01d);
+        assertEquals(95.0, RasterUtils.getNoDataValue(normalizedRaster4.getSampleDimension(1)), 0.01d);
+
+        assertEquals(Arrays.toString(expected3), Arrays.toString(MapAlgebra.bandAsArray(normalizedRaster3, 1)));
+        assertEquals(Arrays.toString(expected4), Arrays.toString(MapAlgebra.bandAsArray(normalizedRaster4, 1)));
+        assertEquals(Arrays.toString(expected5), Arrays.toString(MapAlgebra.bandAsArray(normalizedRaster4, 2)));
+        assertEquals(Arrays.toString(expected7), Arrays.toString(MapAlgebra.bandAsArray(normalizedRaster8, 1)));
+        assertEquals(Arrays.toString(expected8), Arrays.toString(MapAlgebra.bandAsArray(normalizedRaster8, 2)));
+        assertEquals(Arrays.toString(expected9), Arrays.toString(MapAlgebra.bandAsArray(normalizedRaster10, 1)));
+        assertEquals(Arrays.toString(expected10), Arrays.toString(MapAlgebra.bandAsArray(normalizedRaster10, 2)));
+    }
+
+    @Test
+    public void testNormalizeAll2() throws FactoryException {
+        String[] pixelTypes = {"B", "I", "S", "US", "F", "D"}; // Byte, Integer, Short, Unsigned Short, Float, Double
+        for (String pixelType : pixelTypes) {
+            testNormalizeAll2(10, 10, pixelType);
+        }
+    }
+
+    private void testNormalizeAll2(int width, int height, String pixelType) throws FactoryException {
+        // Create an empty raster with the specified pixel type
+        GridCoverage2D raster = RasterConstructors.makeEmptyRaster(1, pixelType, width, height, 10, 20, 1);
+
+        // Fill raster
+        double[] bandValues = new double[width * height];
+        for (int i = 0; i < bandValues.length; i++) {
+            bandValues[i] = i;
+        }
+        raster = MapAlgebra.addBandFromArray(raster, bandValues, 1);
+
+        GridCoverage2D normalizedRaster = RasterEditors.normalizeAll(raster, 0, 255);
+
+        // Check the normalized values and data type
+        double[] normalizedBandValues = MapAlgebra.bandAsArray(normalizedRaster, 1);
+        for (int i = 0; i < bandValues.length; i++) {
+            double expected = (bandValues[i] - 0) * (254 - 0) / (99 - 0);
+            double actual = normalizedBandValues[i];
+            switch (normalizedRaster.getRenderedImage().getSampleModel().getDataType()) {
+                case DataBuffer.TYPE_BYTE:
+                case DataBuffer.TYPE_SHORT:
+                case DataBuffer.TYPE_USHORT:
+                case DataBuffer.TYPE_INT:
+                    assertEquals((int) expected, (int) actual);
+                    break;
+                default:
+                    assertEquals(expected, actual, 0.01);
+            }
+        }
+
+        // Assert the data type remains as expected
+        int resultDataType = normalizedRaster.getRenderedImage().getSampleModel().getDataType();
+        int expectedDataType = RasterUtils.getDataTypeCode(pixelType);
+        assertEquals(expectedDataType, resultDataType);
     }
 
 }
